requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai==1.3.5
openai-whisper==20231117
PyAudio==0.2.14
pydantic==2.5.2
pydantic_core==2.14.5
python-vlc==3.0.20123
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
sympy==1.12
tiktoken==0.5.1
torch==2.1.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
